Revert contents of VOLUME after each RUN
Use DownloadFrom and UploadTo container on the contents of each volume
before and after RUN to preserve the underlying contents of the files.
Due to a limitation in the Docker API, the owners of extracted files are
not preserved. Fixing that will require a side container that
contains tar and can run as root.

Note that this also allows the image to attack the temporary filesystem
of the caller by creating very large VOLUMEs interspersed with RUN
calls.