Check if COPY/ADD destination is a directory
When a user specifies `COPY a /usr/bin` where a is a file, even though
the Dockerfile docs say that this should be a rename, the Docker
1.12/1.13 daemons first check whether the target is a directory before
deciding whether to rename it.

We perform that check by downloading the destination from the container
and checking whether the first entry is a file or a directory. If the
first entry is not found, the location is assumed not to exist and the
target is considered to be a file destination.