Detect and decode errors in progress stream.
This partially reverts commit a87373954e7f63241f7b653471c732adcd9e50ce.

The error is now detected, and passed to pipeIn.CloseWithError,
but that is not generally sufficient to guarantee the writer
(who currently typically ignores errors from imageProgressWriter.Close,
and may not make any further Write calls) will notice.

This, in theory, changes behavior: imageProgressWriter.readProgress
stops reading after detecting a first error and closes the pipe,
possibly causing the writer to abort as well (i.e., typically,
to close the Docker progress connection, possibly canceling the
operation, if anything is still in progress after the error report).

But, AFAICS, the Docker daemon only ever reports an error as the last
item, so in practice this should be mostly invisible.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>