Split an if statement into two, and use pullErr.
We will want to move the call into a defer scope, this will
make it more readable.

Also, use a specially named pullErr variable, to make it clear
that the value must survive the defer scope as well, and avoid
breakage by accidentally defining "err" in new code.

Should not change behavior.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>