fork github.com/docker/docker/builder/dockerfile
The package has been changed upstream quite a bit.  As imagebuilder
relies on the old API, it is incompatible with newer versions of
Moby.  This enforces users of imagebuilder to use an outdated version
of Moby.

Fork github.com/docker/docker/builder/dockerfile to overcome the
incompatibility issue and to allow us to fix potential issues
independently.

All code has been copied from github.docker/docker, so we need to
mention that.  As the files don't have license headers, adding a
NOTICE files seems like a suitable alternative.

Fixes: #116
Signed-off-by: Valentin Rothberg <rothberg@redhat.com>