COPY content of directory & fix '/' in source path
Close #139
When COPY multiple files, check if the destination is a directory. If the destinations is not a directory, stop copying. When the source is a directory, copy the contents of the directory to the destination instead of creating the directory.

treat '/' as current workdir.

compatiable with docker
```
$ tree .
.
└── firstdir
    └── seconddir
        ├── dir-a
        │   └── file-a.txt
        └── dir-b
            └── file-b.txt

FROM alpine:3.10
COPY /firstdir/seconddir/ /var
RUN ls -la /var
/var/dir-a/file-a
/var/dir-b/file-b

FROM alpine:3.10
COPY / /var
RUN ls -la /var
/var/firstdir
```

Signed-off-by: Qi Wang <qiwan@redhat.com>