dockerclient: detect ENOENT when copying
Detect cases where, when filtering the output of archiveFromFile() or
archiveFromContainer() to net only sources which match a source path, we
fail to match anything.  When that happens, return an os.ErrNotExist
error from the archive's Close() method.  The archiveFromContainer()
function now returns a ReadCloser where it previously returned only a
Reader.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>