NewBuilder(): always initialize Args
We modify the Args map member of a Builder object, but we don't make a
copy of the initial value that we're given in NewBuilder() before doing
that, so we end up modifying the map that we're given.  This can be
surprising.

In the dispatcher for ARG instructions, we weren't checking that the map
wasn't nil before attempting to assign values to it, so we could panic.

Fix this by making a fresh map in NewBuilder() and initializing it using
the zero or more values from the passed-in map.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>