Allow Args to be used in Variable Evaluation
If the Dockerfile has a `ARG CHOWN_VAL=6173:6173`, then
a `COPY --chown=${CHOWN_VAL} srcFile destFile` will fail.
However if the Dockerfile has a `ENV CHOWN_VAL=6173:6173`
then it succeeds.

The dispatchCopy() and add() functions were only going through the
list of Environment variables to resolve a Variable instead
of the list of Args and Environment variables.

Now add the definitions from the ARG statements to the ones
from the ENV statements and use both when resolving
variables.

At some point I'd like to look into converting the b.Env to
a Map from a slice like b.Args, but didn't want to run this
down now.

Fixes: https://github.com/containers/buildah/issues/2192
       and probably https://github.com/containers/buildah/issues/2345

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>