Handle ARG and ENV with duplicate name on RUN
If a Dockerfile had an ARG and an ENV with the
same name, things become very unpredictable in
the RUN statement.  This change moves the
code created for handling this same issue in
COPY to the the internals.go file so it
can be used by dispatchers.go and builder.go.

Use the makeUserArgs() function to create a list
of variables, giving precedence to any variables
defined by the ENV statement(s).

This issue was first noted in #159 and once vendored,
this will fix: https://github.com/containers/buildah/issues/2372

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>