Adjust argument slice size in builder
A recent update to the argument handling didn't increase
the argument slice to take into the account the values
from b.Args() that were now being added.  The created
unpredicatble results.

After addressing a comment from @nalind, I realized that
the value from b.Arguments should be used in the steps
as it keeps the ordering of when the ARGS and ENV were
defined in the Dockerfile.  Changed makeUserArgs
in internals.go to take two variables rather than just
a builder.

Adddresses the Buildah bug: https://github.com/containers/buildah/issues/2345

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>