Flag errors on more quoting errors
When parsing a single-quoted token, if we reach an EOF before we reach
the closing quote, flag it as an error.
When parsing a double-quoted token, accept escaped escape characters
("\\"), and if we reach an EOF before we reach the closing quote, flag
it as an error.
This should bring our parsing in line with docker build 17.06 and later.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>