dockerclient.ClientExecutor.Stages(): commit base stages
When we attempt to use the result of a previous stage as a base image
for a later stage, instead of just continuing to use the previous
stage's container, commit the previous stage's container to create an
image which we'll then use to create a new container for use in the
stage.

Builds which looked like this:

  FROM $ExternalBaseImage AS commonBase
  (install general updates)
  FROM commonBase AS builder
  (install build tooling and compile code)
  FROM commonBase AS final
  (copy compiled code from "builder" stage)

would otherwise end up producing an image from its final stage which
contained all of the changes made in the middle stage, which is not
compatible with what `docker build` does.

Note that dockerclient.ClientExecutor.Build() now saves the committed
image's ID in its new Committed.ID field rather than the Image.ID field,
which previously either contained either the ID of a base image or of
the committed image.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>