dockerclient COPY --from: go back and copy skipped hard link targets
If we're copying from another stage or from an archive, and we encounter
an entry that we need to copy that should be hard linked to an entry
that we would have skipped, run through another copy of the archive,
convert the hard link entry we're processing into a regular file entry,
and use the contents of the skipped item as the contents for the regular
file entry.  Rewrite any subsequent links to the skipped entry into
links to the first link, which we converted into the regular file.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>