Avoid panic when writing UIDs/GIDs > 0x1fffff
The USTAR archive format can't encode UIDs or GIDs that can't be
represented in 21 (7 * 3) bits, which creates errors when we modify
headers to include such values.  When we do that to headers that are in
that format, try to force a conversion to the PAX format, which can handle
that.

This requires Go 1.10 or later, as the constants which need to be set in
the header to indicate the format which should be used weren't exported
before that.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>