COPY --chown: add missing parent directories
When we let the daemon implicitly create not-yet-created parent
directories when we COPY in content that is to be owned by a non-root
user, the daemon doesn't know that newly-created directories also need
to be owned by that non-root user, so it creates them owned by root,
which is wrong.

When we're using COPY with --chown, check for the destination path for
any copy being missing, and add any missing path components as
directories.  If the path is a destination file, it'll be overwritten by
the file.  This is pretty expensive, though, so don't do it unless we
have to.

In TestConformanceInternal/copy chown, test --chown with a value larger
than we can represent in 21 bits, so that forceHeaderFormat() will be
invoked.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>