Use Travis for conformance tests
Use Travis for conformance tests, instead of attempting to use GitHub
Actions, which isn't enabled for this repository.

Run unit tests from "make test" using ./... as the package list; the
dependency import tools are smart enough now to not pull in tests for
our dependencies, so we don't need to try to grep out vendored packages.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>