Bump to Go 1.19
Use Go 1.19 instead of 1.16 in go.mod.
Test with 1.19/1.20 in Travis instead of 1.17/1.18.
Bump the Travis environment from the default (16.04) to jammy (22.04),
which is the latest currently available in Travis.
Update assorted direct dependencies.

Update conformance tests nonroot-USER-before-WORKDIR-notused and
nonroot-USER-before-WORKDIR-used to request that they be built with
BuildKit.  Many conformance tests fail if we make this the default;
we'll have to make that change and fix those later.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>