dockerclient: use /bin/sleep for idling instead of sleep
The built-in "sleep" in the latest alpine base image's shell behaves
differently from the version it has to exec(), exiting when it receives
a stray SIGCHLD for one of the processes we're running in an exec
session, and this is causing a consistent failure in the "copy and env
interaction" conformance test.  Call /bin/sleep explicitly to work
around this.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>