Stages.ByName/ByTarget/ThroughTarget: accept numbers
Accept integer values in ByName(), ByTarget(), and ThroughTarget() as
aliases for the stages with that "Position" value.  Add a ByPosition()
alongside ByName().

It would be faster to just index the stages directly instead of
searching for a stage with the right Position value, but this lets us
avoid a bounds check and is just easier to explain.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>